Refactored code into array, record and primitive, and replaced gson with jackson
