+++ maven-plugin/src/main/java/io/confluent/kafka/schemaregistry/maven/derive/schema/protobuf/MapAndArray.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.schemaregistry.maven.derive.schema.utils;
+package io.confluent.kafka.schemaregistry.maven.derive.schema.protobuf;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.ArrayList;
